+++ Tests/Remove-RubrikUser.Tests.ps1
@@ -1,0 +1,40 @@
+Remove-Module -Name 'Rubrik' -ErrorAction 'SilentlyContinue'
+Import-Module -Name './Rubrik/Rubrik.psd1' -Force
+
+foreach ( $privateFunctionFilePath in ( Get-ChildItem -Path './Rubrik/Private' | Where-Object extension -eq '.ps1').FullName  ) {
+    . $privateFunctionFilePath
+}
+
+Describe -Name 'Public/Remove-RubrikUser' -Tag 'Public', 'Remove-RubrikUser' -Fixture {
+    #region init
+    $global:rubrikConnection = @{
+        id      = 'test-id'
+        userId  = 'test-userId'
+        token   = 'test-token'
+        server  = 'test-server'
+        header  = @{ 'Authorization' = 'Bearer test-authorization' }
+        time    = (Get-Date)
+        api     = 'v1'
+        version = '4.0.5'
+    }
+    #endregion
+
+    Context -Name 'Returned Results' {
+        Mock -CommandName Test-RubrikConnection -Verifiable -ModuleName 'Rubrik' -MockWith {}
+        Mock -CommandName Submit-Request -Verifiable -ModuleName 'Rubrik' { return 'OK'}
+        It -Name 'User removed' -Test {
+             Remove-RubrikUser -Id 'test' |
+                Should -BeExactly 'OK'
+        } 
+   
+        Assert-VerifiableMock
+        Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Times 1
+        Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik' -Times 1
+    }
+    Context -Name 'Parameter Validation' {
+        It -Name 'ID Missing' -Test {
+            { Remove-RubrikUser -id } |
+                Should -Throw "Missing an argument for parameter 'id'. Specify a parameter of type 'System.String' and try again."
+        }       
+    }
+}